requirements.txt
# Core IRC bot dependencies
asyncio-mqtt==0.16.1
python-dotenv==1.0.0

# AI and API integrations
google-generativeai==0.3.2
requests==2.31.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Game and utility libraries
astral==3.2
opencv-python==4.8.1.78
mcstatus==11.0.0

# Hardware interfaces (mock for non-RPi systems)
gpiozero==1.6.2

# Financial/trading data  
lightstreamer-client-lib==2.2.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.6

# Database
aiosqlite==0.19.0

# Configuration validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
click==8.1.7
colorama==0.4.6
rich==13.7.0

# Optional: For containerization
gunicorn==21.2.0